Move applyLicenseTemplate to common class
