+++ src/main/kotlin/platform/BaseTemplate.kt
@@ -10,10 +10,8 @@
 
 package com.demonwav.mcdev.platform
 
-import com.demonwav.mcdev.util.License
 import com.intellij.ide.fileTemplates.FileTemplateManager
 import com.intellij.openapi.project.Project
-import java.time.ZonedDateTime
 
 abstract class BaseTemplate {
 
@@ -28,17 +26,5 @@ abstract class BaseTemplate {
         properties?.let { prop -> allProperties.putAll(prop) }
 
         return template.getText(allProperties)
-    }
-
-    fun applyLicenseTemplate(
-        project: Project,
-        license: License,
-        author: String
-    ): String {
-        val props = mapOf(
-            "YEAR" to ZonedDateTime.now().year.toString(),
-            "AUTHOR" to author
-        )
-        return project.applyTemplate("${license.id}.txt", props)
     }
 }
